src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==25.1.0",
    "aiosmtplib==5.0.0",
    "alembic==1.17.2",
    "amqp==5.3.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "apscheduler>=3.11.1",
    "asttokens==3.0.1",
    "asyncpg==0.30.0",
    "bcrypt>=4.2.0,<5.0.0",
    "billiard==4.2.3",
    "black==25.11.0",
    "ldap3==2.9.1",
    "celery==5.5.3",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "click==8.3.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "coverage==7.11.3",
    "cryptography==46.0.3",
    "deprecated==1.3.1",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "executing==2.2.1",
    "fastapi==0.121.2",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.3.1",
    "greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "kombu==5.5.4",
    "limits==5.6.0",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "minio>=7.2.20",
    "nh3>=0.2.0",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "passlib>=1.7.4",
    "pathspec==0.12.1",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "prometheus-client==0.23.1",
    "prometheus-fastapi-instrumentator==7.0.0",
    "prompt-toolkit==3.0.52",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.1",
    "pycparser==2.23 ; implementation_name != 'PyPy'",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyspnego==0.12.0",
    "pytest==9.0.1",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    # REMOVED: "python-jose==3.5.0" - CVE-2022-29217, migrated to pyjwt
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "pytokens==0.3.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==7.1.0",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.6",
    "rsa==4.9.1",
    "ruff==0.14.5",
    "sentry-sdk==2.44.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slowapi==0.1.9",
    "smbprotocol==1.15.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "sqlmodel==0.0.27",
    "sspilib==0.4.0 ; sys_platform == 'win32'",
    "starlette==0.49.3",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'",
    "vine==5.1.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websockets==13.1",
    "wrapt==2.0.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
]
